// src/models/Soap/client-info.ts
export interface ClientInfo {
  WSBCC: {
    ENCABEZADO: Encabezado
    DETALLE: {
      TIPODOCCLI: Text
      DOCCLI: Text
      FECHAEXPDOC: Text
      NOMBRES1: Text
      NOMBRES2: Text
      APELLIDO1: Text
      APELLIDO2: Text
      TIPOPERS: Text
      TIPOCLI: Text
      RETFUENTE: Text
      DIRRES: Text
      TELEFONO: Text
      CIUDADRES: Text
      ORGPUBLICO: Text
      NOMINA: Text
      FALLECIDO: Text
      FECFALLEC: Text
      SEXO: Text
      COOPPER: Text
      JCOMUNPER: Text
      UBICACPER: Text
      ESTRATOPER: Text
      CIUDADEXPDOC: Text
      CIUDADNAC: Text
      PAISNAC: Text
      FECHANAC: Text
      ESTCIVIL: Text
      NIVEST: Text
      PROFESION: Text
      BARRIORES: Text
      DEPRES: Text
      CELULAR: Text
      NOMEMPLAB: Text
      CARGO: Text
      DIREMPLAB: Text
      TELFEMPLAB: Text
      CIUDADEMPLAB: Text
      NOMBREREFFAM: Text
      PARENREFFAM: Text
      DIRREFFAM: Text
      TELFREFFAM: Text
      NOMBREREFPER: Text
      DIRREFPER: Text
      TELFREFPER: Text
      NOMBREREFCOM1: Text
      DIRREFCOM1: Text
      TELFREFCOM1: Text
      ORIGENFONDOS: Text
      INGRESOS: Text
      EGRESOS: Text
      ACTIVOS: Text
      PASIVOS: Text
      OTROSINGRESOS: Text
      CIIU: Text
      TIPOSOCIEDAD: Text
      TIPOACCASO: Text
      FECVINC: Text
      TIPOVIVIENDA: Text
      OCUPACION: Text
      DESCOCUPA: Text
      TIPOEMPLAB: Text
      OPERINTER: Text
      MONINTER: Text
      TIPOOPERINTER: Text
      OTROTIPOOPERINTER: Text
      NOMENTOPERINT: Text
      NUMCTAOPERINT: Text
      CIUDADOPERINT: Text
      PAISOPERINT: Text
      MONOPERINT: Text
      TIPOINM1: Text
      DIRINM1: Text
      CIUDADINM1: Text
      VALCOMINM1: Text
      SALDOHIPINM1: Text
      ENTPERINM1: Text
      TIPOINM2: Text
      DIRINM2: Text
      CIUDADINM2: Text
      VALCOMINM2: Text
      SALDOHIPINM2: Text
      ENTPERINM2: Text
      MARCAVEH1: Text
      MODELOVEH1: Text
      PLACAVEH1: Text
      VALORCOMVEH1: Text
      SALDOPIGVEH1: Text
      ENTPERVEH1: Text
      MARCA_VEH2: Text
      MODELOVEH2: Text
      PLACAVEH2: Text
      VALORCOMVEH2: Text
      SALDOPIGVEH2: Text
      ENTPERVEH2: Text
      MONITOREO: Text
      PEPS: Text
      DOCTUTOR: Text
      TIPODOCTUTOR: Text
      NOMBRESTUTOR: Text
      MONTONEGINT: Text
      DIRENVCORRESP: Text
      EXTTELF: Text
      DESOTROSING: Text
      EMAIL: Text
      AUTAENVSMS: Text
      AUTENVEMAIL: Text
      DECLARANTE: Text
      ENVCERTCOSTOS: Text
      ENVEXTRACTOS: Text
      INCAPACIDADTOTAL: Text
      CIUDADCORR: Text
      PERMFATCA: Text
      NACAMERICANA: Text
      RESEEUU: Text
      RELCONTRACTUAL: Text
      FAMPEPSAPELLIDO1: Text
      FAMPEPSAPELLIDO2: Text
      FAMPEPSNOMBRES: Text
      FAMPEPSPAREN: Text
      CODDISCAPACIDAD: Text
      DOCASESORCOM: Text
      SECTOREMPRESA: Text
      'REF-FIN-PEPS': Text
      'DIR-FIN-PEPS': Text
      'CIU-FIN-PEPS': Text
      'TEL-FIN-PEPS': Text
      'FEC-EST-FINAN': Text
      'ORIGEN-EST-FINAN': Text
      SIGLA: Text
      NOMBREJURIDICO: Text
      TIPOEMPJUR: Text
      EMPFAX: Text
      EMAILENVIO: Text
      TIPOPERJUR: Text
      MARCALAFT: Text
      INDSUC: Text
      CIUDADSUC: Text
      SUCDIR: Text
      SUCTEL: Text
      SUCFAX: Text
      RF1INDJUR: Text
      RF1ENTJUR: Text
      RF1PROJUR: Text
      RF1CTAJUR: Text
      RF1DIRJUR: Text
      RF2INDJUR: Text
      RF2ENTJUR: Text
      RF2PROJUR: Text
      RF2CTAJUR: Text
      RF2DIRJUR: Text
      RC1NOMJUR: Text
      RC1CIUJUR: Text
      RC1TELCEL: Text
      RC2INDJUR: Text
      RC2NOMJUR: Text
      RC2CIUJUR: Text
      RC2TELCEL: Text
      ORIGFONDOSJUR: Text
      RLTIPODOC: Text
      RLCED: Text
      RLPRIAPE: Text
      RLSEGAPE: Text
      RLNOMBRES: Text
      RLEXPEDIDA: Text
      RLNACIMIENTO: Text
      RLFECNACIMIE: Text
      RLFECHAEXPDOC: Text
      RLINGRESOSMENS: Text
      RLVLREGRESOS: Text
      RLVLROTROSIN: Text
      RLVLRTOTACTI: Text
      RLVLRTOTPASI: Text
      RLCONCEPTOINGRE: Text
      RLEMAIL: Text
      RLDIR: Text
      RLCIU: Text
      RLTEL: Text
      RLPEPS: Text
      RLNOMESTPEPS: Text
      RLDIRESTPEPS: Text
      RLCIUESTPEPS: Text
      RLTELESTPEPS: Text
      RLRPIND: Text
      RLRPNOMBRES: Text
      RLRPCIUDAD: Text
      RLRPTEL: Text
      RLOPERINT: Text
      RLMONINTE: Text
      RLTOPEINT: Text
      RLDESOTRO: Text
      RLNBCOINT: Text
      RLNCTAINT: Text
      RLCIUDINT: Text
      RLPAISINT: Text
      RLMONEINT: Text
      RLMONTINT: Text
      RLSUPIND: Text
      RLSUPTIPODOC: Text
      RLSUPCED: Text
      RLSUPPAPE: Text
      RLSUPSAPE: Text
      RLSUPNOMB: Text
      INDTINFATCA: Text
      NUMTINFATCA: Text
      PARTICFATCA: Text
      CLIESPECIAL: Text
      ACCNIVEL2: Text
      ACCNIVEL3: Text
      MIEMBROSJUNTA: Text
    }
  }
}